fixed version

+++ lab2.java
@@ -4,10 +4,10 @@ public class lab2 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         int temp = 1;
-        int factorial = 1;
+        long factorial = 1;
         System.out.print("Введите число от 1 до 15 - ");
         if (scanner.hasNextInt()) {
-        int chislo = scanner.nextInt();
+            int chislo = scanner.nextInt();
             if (chislo <= 15 && chislo >= 0) {
                 while (temp <= chislo) {
                     factorial = factorial * temp;
